refactor: rename :baselineprofile module to :benchmarks
Aligns with community convention (NIA: benchmarks/, Pokedex: benchmark/).
Module contained both baseline profile generator and macrobenchmarks —
the old name only reflected half the purpose.

Updates: settings.gradle.kts, app/build.gradle.kts, namespace,
package declarations (de.lemke.geticon.benchmarks), CLAUDE.md.
