+++ benchmarks/src/main/AndroidManifest.xml
